Add Jupyter notebook summarization features
This update introduces a summarization function for Jupyter notebooks. The new feature enables parsing and summarizing of cells from a provided Jupyter notebook. The summary is generated by OpenAI's GPT-3 model. We also have included rouge_score and BLEU score functions for evaluation of the summary. Moreover, Streamlit is used to create a web app that accepts Jupyter files, computes summary and evaluates it via rouge and BLEU scores. The summarized version of the notebook can be downloaded via a provided link. Corresponding modules were added to the requirements file.